+++ src/commands/update_node_provider.rs
@@ -27,7 +27,7 @@ pub fn exec(
     auth: &AuthInfo,
     opts: UpdateNodeProviderOpts,
 ) -> AnyhowResult<Vec<IngressWithRequestId>> {
-    let reward_account = ledger_canister::AccountIdentifier::from_hex(&opts.reward_account)
+    let reward_account = icp_ledger::AccountIdentifier::from_hex(&opts.reward_account)
         .map_err(|e| anyhow!(e))
         .with_context(|| format!("Account {} is not valid address", &opts.reward_account))?;
     let args = Encode!(&UpdateNodeProvider {