+++ CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Changed
+
 - remove the `EC PARAMETERS` section in the PEM file to match dfx
+
+### Added
+
+- ckBTC commands and support. (#153)
 
 ## [0.3.2] - 2023-01-13